fix: reset runTests store after event handling
Fixes the issue described in #2 where the app bar button for running all
tests doesn't react to clicks if a test run fails.

The root cause was that the runTestsEvent event subscriber in the app layout
would not be called by subsequent store updates as the value would be set to
`./...` which is the value already stored, so update subscribers were
not called.

The issue is fixed by making the layout runTestsEvent subscriber reset
the store value to an empty string after being called.