fix: reset openSettingsEvent store after each event handling
The store subscription responsible for opening the settings drawer did
not reset the store's value after acting on the change, which caused
settings drawer to not open after being opened the first time.

The value is now set to `null` after each handling to ensure the
subscription is triggered again when the store is set to a boolean
value.

Fixes #7